+++ src/algorithm/population/PoblacionInterface.java
@@ -6,6 +6,6 @@ import algorithm.functions.Function;
 import algorithm.individuos.Individuo;
 
 public interface PoblacionInterface {
-    public List<Individuo> generaPoblacion(String type, int depth, int size, Function function);
+    public List<Individuo> generaPoblacion(String type, double depth, int size, Function function);
     
 }
